Better symlink race protection
We want to support symlinks so tools can put their service
manifest in their git repo and put it anywhere, but also want
to make sure that they don't allow us to read files that belong
to someone else. This is a compromise fix that requires files
to be owned by the tool's uid that should work, I guess.

Bug: T95210
Change-Id: Icb9bbbcf27d555af00b7bca798ac43c7af45d91c