toolforge: Increase WebServiceMonitor sleep time to 60s
Considering our NFS performance problems plus the fact we don't
have formal SLAs, it seemes a 10s reaction time is too aggressive
for monitoring.

If faster reaction times are required, deployment on Kubernetes
should be considered (althouh the deployment controller probably
won't be that fast either).

Additionally, this will improve load on the LDAP server as well
as avoid a flood of webservice processes spawning and killing
everything if the cluster goes down for any reason.

Bug: T210190
Change-Id: I78943fb1c2a5f283cb0dbbcd70766953e4f58371