Add `--backend=gridengine` to restart command
This works around a bug in `webservice` that is keeping it from properly
detecting the backend from a $HOME/service.manifest file. That bug
should be fixed, but this workaround is also safe as we are already
asserting that $HOME/service.manifest declares the desired backend is
gridengine before looking for a running job and eventually calling for
a restart.

Bug: T263190
Change-Id: I852320be14ff0e50efc3740cb4118c460aa855e9