debian: drop help2man
The man pages created by help2man are totally broken as the libraries
the program uses are not present during build. Completely dropping the
man page is likely the best option; "collector-runner --help" is not
particularly helpful to Toolforge roots debugging issues with it and our
time is better spent completely replacing the grid.

Change-Id: I46294f94bcf15d3d4762bd447c5c0f7e2838bb1d